Add timeout to md-scroll-spy
The time in ms to wait before computing links. Useful for animations
that affect visibility. Set the timeout to be longer than the animation
to make sure rendering is complete.

If we ever remove the dependency on jQuery for animations, we can
probably find a better option than a timeout. This will mostly be
accomplished by #314.